Add Terrain Tools 3.0.2-preview.3 as a local package
This is the specific version of this package that Ryan Zehm's importer is written for, and the API is so different between this old version and the newer, stable versions that there's no way I'm going to try and update this importer to work with those, at least not for the time being.

Note: All project errors from the previous commit should be gone now.